requirements.txt
numpy==1.24.1
plotly==5.13.0
opencv-python-headless==4.7.0.68
matplotlib==3.6.3
torch==1.13.1
Pillow==9.4.0
scipy==1.10.0
audioplayer